cn0326: add cn0326 demo project
This patch adds support for cn-0326 Isolated Low Power pH Monitor with
Temperature Compensation.

The board displays a prompt message on the terminal and a list of commands
is accessible via the 'help' comand.
It is possible to use cn-0326 and adicup3029 to read pH data via terminal.

Refrence Design
http://www.analog.com/media/en/reference-design-documentation/reference-designs/CN0326.pdf