cn0326: fixed bug with the SPI communication introduced at porting
The code for adicup360 used a GPIO to control the CS pin of the SPI. As
adicup3029 does not use a GPIO the code was commented and the if statement
started to guard the while statement, that neede to happen regardles of
if. The if was removed as it was redundant and the bug is thusly removed.

Signed-off-by: adrimbarean <Andrei.Drimbarean@analog.com>