cn0415: initial commit
The EVAL-CN0415-ARDZ is a robust pulse-width modulation (PWM) solenoid
driver circuit in an Arduino shield form factor.

https://wiki.analog.com/resources/eval/user-guides/eval-adicup3029/hardware/cn0415

This application uses the board with the EVAL-ADICUP3029 to implement a
PID controller designed to follow a current target in the lead or a %
duty cycle of the PWM. It has dither capabilities and over-voltage and
over-current protections. This commit adds the CCES project files and
the main and configuration files of the project.

More details on the application at:
https://wiki.analog.com/resources/eval/user-guides/eval-adicup3029/reference_designs/demo_cn0415

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>