cli_demo: fix compiler path to "include" folder
The compiler path to the "include" folder was a literal path to the
folder three on my specific machine and not a variable path to the
workspace folder. As such, the project could not be compiled on other
machines. Update the compiler path to the "include" folder to resolve
this issue.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>