cn0414: add AD5700 HART modem implementation
The AD5700 in a HART modem that uses an extended UART interface to drive
HART messages. The extended pins are Request To Send(nRTS) and (Carrier
Detect)CD. The driver can receive messages asynchoniously through GPIO
interrupt on CD pin. The AD5700_gpio modules are to provide abstraction
between the GPIO platform drivers and the AD5700 module and keep the
layered approach of the program.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>