ADuCM3029_demo_cn0397: Add main application
This patch adds the application source. It is modified to use the new BLE
message format intended to work with IOTnode IOS app.

This is just a port of the ADICUP3029 BSP cn0397 project with
modifications made to the BLE message format.

For information about the BLE message format check:
https://wiki.analog.com/resources/eval/user-guides/eval-adicup3029/reference_designs/ble_packet

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>