cn0418: add project implementation
The ADuCM3029_demo_cn0418 project provides a solution to control a PLC
or DCS output system using the EVAL-CN0418-ARDZ and the EVAL-ADICUP3029.
It uses a DAC with 4 channels that can be configured as either voltage
or current output and is compatible with HART modem output to enable
HART communication on any channel. It has a 32kb EEPROM memory that can
also be used to identify the board and is controlled via a command line
interface (CLI).

More information at:
https://wiki.analog.com/resources/eval/user-guides/eval-adicup3029/reference_designs/demo_cn0418

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>