cn0435: add exiting project components
The project is a dynamic PLC/DCS system that includes two types of boards:

 - EVAL-CN0414-ARDZ:
   https://wiki.analog.com/resources/eval/user-guides/eval-adicup3029/hardware/cn0414
 - EVAL-CN0418-ARDZ:
   https://wiki.analog.com/resources/eval/user-guides/eval-adicup3029/hardware/cn0418

Files that make the demo for those projects are used in this project as
well. They are added in this commit unmodified, but arranged so that files
common to both projects are located in the root source folder and files
that are specific to one or the other projects are in separate child
folders.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>