cn0435: add CLI module
The system works in two modes of opperation: CLI and MODBUS. Both are
characterized by the way it communicates with a master control device: the
PC in CLI mode and a MODBUS master in MODBUS mode.

The CLI mode of opperation provides a debug method for the system. Add the
CLI module to the application and a simple configuration file used to
select the mode of operation.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>