cn0435: add MODBUS module
The system works in two modes of opperation: CLI and MODBUS. Both are
characterized by the way it communicates with a master control device: the
PC in CLI mode and a MODBUS master in MODBUS mode.

The MODBUS mode of operation is intended to be the main mode of operation
for this application. The system is a MODBUS slave and responds to a
master's commands. Add the MODBUS slave module. This is the data link layer
of the MODBUS application.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>