cn0435: modify CN0414 and CN0418 application code
Update the CN0414 and CN0418 application modules to make them extensible in
this context. To work together each application will only instantiate
handlers that are needed for each instance individually and leave the rest
of the component drivers to be instantiated by the system manager module.

The CN0414 will only instantiate the AD4111 ADC driver and the
adc_update_timer module. The latter only in CLI mode.

The CN0418 will only instantiate the AD5755 DAC driver.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>