projects:no-os_library_project: Add project to build no-os
This projecte builds the noos files as a static library and it can be used by any adicup project.
Adicup repository projecte will be directly linked to it and just have to be imported.

If a new project wants to use it, it should be imported as a link in the developer workspace,
the new project should mark no-os project as a reference and add include dependecies from it
(no-os/includes, no-os/platform_drivers)

More details will be described in a wiki page.

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>