Build_projects.yml: Create file for pipeline to build projects
A build will be triggered for new commits on master or dev/*
branches an on PR to main.
A self hosted pool will be used to run this pipelene.
Binaries will be build by calling build_projects.py.
Output files will be exported as an artifcat.
Variables as RELEASE_DIR and ARTIFACT_NAME are defined
in the variable group noos_releases_group_variables inside Azure in
order to be shared with the release pipeline

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>