Add fluorescence decay support to the cn0503 firmware.
Add four commands to the cn0503 CLI: IMPRESP, FLUO_CALIB, FLUO_DECAY,
FLUO_DUMP_CALIB.

For all commands, information pertaining to impulse response
measurements is stored in the device structure, as a struct
cn0503_impulse_response *. See cn0503.h for the definition of this
struct.

IMPRESP (handler function: cn0503_impulse_response): measures the
impulse response of a channel by performing an initial LED pulse and
then taking samples over a period of time after the end of the pulse.
This is achieved one of two ways. We can use impulse response mode,
which takes consecutive ADC samples after a single LED pulse. Or, we can
use normal integration mode (with the integrator configured as
single-sided, or as a buffer), and increment the integration offset
after each sample.

Helper functions:
- impulse_response_rollback: frees all memory malloc'd during an impulse
  response measurement, and rolls back any register configuration changes
  to their previous values.
- impulse_response_start_collection: clears the FIFO. Sets the
  integration offset if needed. Starts data collection.
- impulse_response_read_data: reads data from the FIFO. Outputs data
  through the CLI or repeats sampling process.
- impulse_response_fifo_callback: Sets a flag in the device structure
  that it is ready to read data.
- impulse_response_parse_arg: parse a single command-line argument for
  the impulse response, fluorescence calibration, or fluorescence decay
  commands.
- impulse_response_parse_arg_list: parse a list of arguments for the
  IMPRESP, FLUO_CALIB, or FLUO_DECAY commands.
- impulse_response_validate_parameters: makes sure the parameters for an
  impulse response measurement are valid (eg., the sampling period
  specified is not too small, etc.)
- impulse_response_setup: sets the register configuration for the
  impulse response measurement. Stores previous register values so we
  can roll back our configuration changes at the end.

FLUO_CALIB (handler function: cn0503_fluo_decay_calibrate): measures the
impulse response of a specified optical path, and stores the result as a
calibration curve in flash memory.

Helper functions:
- process_fluo_decay_calibration: writes calibration curve to flash
  memory (to the page specified by CN0503_FLASH_FLUO_CALIB_ADDR in
  cn0503.h, and pages below that address).
- cn0503_fluo_set_presets: FLUO_CALIB and FLUO_DECAY have an option to
  use preset values in lieu of command-line arguments (for parameters
  such as the sampling period, LED pulse width, etc.). This function
  loads these presets into the cn0503_impulse_response structure in the
  device structure.

FLUO_DECAY (handler: cn0503_fluo_decay_measure): Measures the impulse
response of a fluorophore in the specified optical path. Reads the
calibration curve from flash, and subtracts it from the impulse response
of the fluorophore. Performs an exponential fit on the subtracted data.

Helper functions:
- compute_decay_constant: reads the calibration curve from flash and
  subtracts it from the fluorescence impulse response measured. Performs
  an exponential fit on the resulting data. Outputs the exponential fit,
  subtracted data, and average squared error of the fit to the data
  through the CLI.

FLUO_DUMP_CALIB (handler: cn0503_fluo_dump_calib): reads the calibration
curve from flash and outputs it to the CLI.

Bug fixes to flash.c.

In places, the flash address was treated as the address of a
byte-indexed memory, and in others it was treated as the address of a
word-indexed memory (flash on the ADICUP3209 is byte-indexed).