cn0567: initial commit
Initial version of the CN0567 project.

Provides IIO support for both versions of the reference design:
ADPD4100 and ADPD4101.

A part of the requirement is to have the device configuration separate
from the application implementation for easier usage by the customers
(timeslot configuration / register configuration). This is done via
`cn0567_config.h` file.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>