projects/ADuCM3029_demo_cn0536: Update init func
init_communication was trying to start communication
over mqtt and if it wasn't working, fallback to UART.
At the end this make no sens because the UART port
is connected to the ESP8266 module not to the PC,
therefore there is no output in this scenario. Also
if mqtt fails this time, error LED will blink instead
of notify the UART success initialization.

Signed-off-by: Mihail Chindris <mihail.chindris@analog.com>